更改資料傳遞的位置
