+++ src/components/AdminMain/AdminMainTweetList.jsx
@@ -1,40 +1,10 @@
-import { useState } from 'react'
 import styles from './AdminMainTweetList.module.scss'
 import AdminTweetItem from './AdminTweetItem'
 
 
-const tweetList = [
-  {
-    id: 1,
-    date: 36,
-    text: 'nothing',
-    user: {
-      avatar: 'https://play-lh.googleusercontent.com/69DEGMGZWRsrGy6AyIZ3k8xJGMGXHfr6jzn63IBBcyD3CgyixnmRZIfMf0QrbdWUQTBg=w240-h480-rw',
-      userName: 'Allen',
-      account: 'Joe123'
-    }
-  },
-  {
-    id: 2,
-    date: 45,
-    text: 'nothing is over',
-    user: {
-      avatar: 'https://storystudio.tw/storage/upload/article/2020/03/people-x-sobooks-baxuan-2.jpg',
-      userName: 'Allen',
-      account: 'Joe456'
-    }
-  }
-]
 
-const AdminMainTweetList = () => {
-  const [tweets, setTweets] = useState(tweetList)
-
-
-  function handleDelete(id){
-    console.log('delete')
-
-    setTweets((prevTweets) => prevTweets.filter(tweet => tweet.id !== id))
-  }
+const AdminMainTweetList = ({ tweets, onDelete}) => {
+  
 
   return (
     <div className={styles.adminMain}>
@@ -47,7 +17,7 @@ const AdminMainTweetList = () => {
             <AdminTweetItem
               key={tweet.id}
               tweet={tweet}
-              onDelete={handleDelete}
+              onDelete={(id) => onDelete?.(id)}
             />
           )
         })}
